refactor(mesh/cleanup): extract directed_edges and twin_edges primitives
Two related dedups in cleanup/:

- invariants.rs: find_twin_edges and find_unrepaired_tjunctions both
  walked polygon edges as (vertices[i], vertices[(i+1) % n]) pairs
  with the wrap-around modulo open-coded. Hoist that into a private
  directed_edges(poly) iterator in invariants.rs so the indexing
  appears exactly once.

- merge::boundary_edges_after_twin_cancellation and
  provenance::unmatched_edges were structurally identical
  implementations of "for each canonical edge, push surplus copies
  of dominant direction" (provenance.rs's doc even called this out:
  "Same shape as merge::boundary_edges_after_twin_cancellation but
  global"). Hoist to new cleanup/twin_edges.rs as
  surviving_directed_edges; both call sites forward to it.

The merge.rs and provenance.rs functions stay as thin wrappers
because their call sites address the local name + their own doc
context (per-bucket vs whole-mesh) is more useful than collapsing
through to the shared name everywhere.

Closes the DC findings at invariants.rs:146, invariants.rs:257,
merge.rs:212. The test-fixture DC findings (invariants.rs:348,
merge.rs:701, merge.rs:764) stay — each test mesh is intentionally
self-contained so the scenario reads top-to-bottom.

Verified: cargo nextest run -p aether-mesh --lib cleanup
(95/95 pass).
