+++ crates/aether-mesh/src/cleanup/invariants.rs
@@ -17,10 +17,19 @@
 //! Geometric correctness (manifold-ness, no self-intersection) belongs
 //! in a separate validation pass; this module is about pass composition.
 
-use super::mesh::{IndexedMesh, VertexId};
+use super::mesh::{IndexedMesh, IndexedPolygon, VertexId};
 use super::tjunctions::is_strictly_between;
 use crate::plane::Plane3;
 use crate::point::Point3;
+
+/// Iterate `poly`'s directed edges as `(vertices[i], vertices[(i+1) % n])`
+/// pairs in vertex order. Shared by the twin / T-junction invariant
+/// scans so the wrap-around `(i + 1) % n` indexing is written exactly
+/// once.
+fn directed_edges(poly: &IndexedPolygon) -> impl Iterator<Item = (VertexId, VertexId)> + '_ {
+    let n = poly.vertices.len();
+    (0..n).map(move |i| (poly.vertices[i], poly.vertices[(i + 1) % n]))
+}
 
 /// One twin-edge violation surfaced by [`find_twin_edges`].
 #[derive(Debug, Clone)]
@@ -141,11 +150,7 @@ pub fn find_twin_edges(mesh: &IndexedMesh) -> Vec<TwinEdgeViolation> {
     for poly in &mesh.polygons {
         let key = bucket_key(&poly.plane, poly.color);
         let entry = directed.entry(key).or_default();
-        let m = poly.vertices.len();
-        //noinspection DuplicatedCode
-        for i in 0..m {
-            let a = poly.vertices[i];
-            let b = poly.vertices[(i + 1) % m];
+        for (a, b) in directed_edges(poly) {
             *entry.entry((a, b)).or_insert(0) += 1;
         }
     }
@@ -252,11 +257,7 @@ pub fn find_unrepaired_tjunctions(mesh: &IndexedMesh) -> Vec<UnrepairedTJunction
     use std::collections::HashSet;
     let mut edges: HashSet<(VertexId, VertexId)> = HashSet::new();
     for poly in &mesh.polygons {
-        let n = poly.vertices.len();
-        //noinspection DuplicatedCode
-        for i in 0..n {
-            let a = poly.vertices[i];
-            let b = poly.vertices[(i + 1) % n];
+        for (a, b) in directed_edges(poly) {
             if a == b {
                 continue;
             }
